+++ frontend/sw.js
@@ -96,6 +96,7 @@ const SHELL = [
   "./js/install.js",
   "./js/live.js",
   "./js/offline.js",
+  "./js/scrollbar.js",
   "./js/share.js",
   "./js/browserdata.js",
   "./js/components/vote.js",